Move start workflow to api package
