+++ service/history/api/signalwithstartworkflow/signal_with_start_workflow_test.go
@@ -22,7 +22,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package signalwithstart
+package signalwithstartworkflow
 
 import (
 	"context"
